# Improve performance of module dictionary collection